+++ quadrama/main/java/de/unistuttgart/ims/coref/annotator/plugin/quadrama/tei/QuadramaTeiExportPlugin.java
@@ -6,22 +6,16 @@ import javax.swing.filechooser.FileFilter;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.uima.analysis_engine.AnalysisEngineDescription;
-import org.apache.uima.collection.CollectionReaderDescription;
-import org.apache.uima.fit.component.NoOpAnnotator;
-import org.apache.uima.fit.factory.AggregateBuilder;
 import org.apache.uima.fit.factory.AnalysisEngineFactory;
-import org.apache.uima.fit.factory.CollectionReaderFactory;
 import org.apache.uima.resource.ResourceInitializationException;
 
 import de.unistuttgart.ims.coref.annotator.ExtensionFilters;
 import de.unistuttgart.ims.coref.annotator.FileFilters;
-import de.unistuttgart.ims.coref.annotator.plugin.quadrama.QDStylePlugin;
-import de.unistuttgart.ims.coref.annotator.plugins.AbstractIOPlugin;
-import de.unistuttgart.ims.coref.annotator.plugins.IOPlugin;
-import de.unistuttgart.ims.coref.annotator.plugins.StylePlugin;
+import de.unistuttgart.ims.coref.annotator.plugins.AbstractExportPlugin;
+import de.unistuttgart.ims.coref.annotator.plugins.UimaExportPlugin;
 import javafx.stage.FileChooser.ExtensionFilter;
 
-public class Plugin extends AbstractIOPlugin implements IOPlugin {
+public class QuadramaTeiExportPlugin extends AbstractExportPlugin implements UimaExportPlugin {
 
 	@Override
 	public String getDescription() {
@@ -40,32 +34,14 @@ public class Plugin extends AbstractIOPlugin implements IOPlugin {
 	}
 
 	@Override
-	public AnalysisEngineDescription getImporter() throws ResourceInitializationException {
-		AggregateBuilder b = new AggregateBuilder();
-		b.add(AnalysisEngineFactory.createEngineDescription(NoOpAnnotator.class));
-		return b.createAggregateDescription();
-	}
-
-	@Override
 	public AnalysisEngineDescription getExporter() throws ResourceInitializationException {
 		return AnalysisEngineFactory.createEngineDescription(MapCorefToXmlElements.class);
-	}
-
-	@Override
-	public CollectionReaderDescription getReader(File f) throws ResourceInitializationException {
-		return CollectionReaderFactory.createReaderDescription(TeiReader.class, TeiReader.PARAM_SOURCE_LOCATION,
-				f.getAbsoluteFile(), TeiReader.PARAM_LANGUAGE, "de", TeiReader.PARAM_DOCUMENT_ID, f.getName());
 	}
 
 	@Override
 	public AnalysisEngineDescription getWriter(File f) throws ResourceInitializationException {
 		return AnalysisEngineFactory.createEngineDescription(TeiWriter.class, TeiWriter.PARAM_OUTPUT_FILE,
 				f.getAbsolutePath());
-	}
-
-	@Override
-	public Class<? extends StylePlugin> getStylePlugin() {
-		return QDStylePlugin.class;
 	}
 
 	@Override