+++ src/main/java/de/unistuttgart/ims/coref/annotator/Defaults.java
@@ -2,6 +2,7 @@ package de.unistuttgart.ims.coref.annotator;
 
 public class Defaults {
 	public static final String CFG_ANNOTATOR_ID = "Unknown annotator";
+	public static final int CFG_AUTOSAVE_TIMER = 300000;
 	public static final boolean CFG_COMPARE_BY_ENTITY_NAME = true;
 	public static final boolean CFG_CREATE_DEFAULT_FLAGS = false;
 	public static final boolean CFG_DELETE_EMPTY_ENTITIES = false;
@@ -17,5 +18,6 @@ public class Defaults {
 	public static final boolean CFG_ASK_BEFORE_FILE_OVERWRITE = true;
 	public static final boolean CFG_IGNORE_SINGLETONS_WHEN_COMPARING = true;
 	public static final boolean CFG_UNDERLINE_SINGLETONS_IN_GRAY = false;
+	public static final boolean CFG_STICKY_FLAGS = false;
 
 }