+++ src/main/java/de/unistuttgart/ims/coref/annotator/DocumentWindow.java
@@ -91,7 +91,6 @@ import org.apache.uima.fit.util.JCasUtil;
 import org.apache.uima.jcas.JCas;
 import org.apache.uima.jcas.cas.TOP;
 import org.apache.uima.jcas.tcas.Annotation;
-import org.apache.uima.resource.ResourceInitializationException;
 import org.eclipse.collections.api.list.ImmutableList;
 import org.eclipse.collections.api.list.MutableList;
 import org.eclipse.collections.api.set.ImmutableSet;
@@ -107,6 +106,7 @@ import de.unistuttgart.ims.coref.annotator.action.ChangeKeyForEntityAction;
 import de.unistuttgart.ims.coref.annotator.action.CopyAction;
 import de.unistuttgart.ims.coref.annotator.action.DeleteAction;
 import de.unistuttgart.ims.coref.annotator.action.DeleteAllMentionsInSelection;
+import de.unistuttgart.ims.coref.annotator.action.DuplicateMentions;
 import de.unistuttgart.ims.coref.annotator.action.EntityStatisticsAction;
 import de.unistuttgart.ims.coref.annotator.action.ExampleExport;
 import de.unistuttgart.ims.coref.annotator.action.FileExportAction;
@@ -128,6 +128,7 @@ import de.unistuttgart.ims.coref.annotator.action.RenameEntityAction;
 import de.unistuttgart.ims.coref.annotator.action.SelectNextMentionAction;
 import de.unistuttgart.ims.coref.annotator.action.SelectPreviousMentionAction;
 import de.unistuttgart.ims.coref.annotator.action.SetLanguageAction;
+import de.unistuttgart.ims.coref.annotator.action.ShowDocumentStatistics;
 import de.unistuttgart.ims.coref.annotator.action.ShowFlagEditor;
 import de.unistuttgart.ims.coref.annotator.action.ShowLogWindowAction;
 import de.unistuttgart.ims.coref.annotator.action.ShowMentionInTreeAction;
@@ -177,11 +178,13 @@ import de.unistuttgart.ims.coref.annotator.document.op.RemoveMention;
 import de.unistuttgart.ims.coref.annotator.document.op.UpdateEntityName;
 import de.unistuttgart.ims.coref.annotator.plugin.rankings.MatchingRanker;
 import de.unistuttgart.ims.coref.annotator.plugin.rankings.PreceedingRanker;
-import de.unistuttgart.ims.coref.annotator.plugins.DefaultIOPlugin;
+import de.unistuttgart.ims.coref.annotator.plugins.DefaultImportPlugin;
 import de.unistuttgart.ims.coref.annotator.plugins.EntityRankingPlugin;
-import de.unistuttgart.ims.coref.annotator.plugins.IOPlugin;
+import de.unistuttgart.ims.coref.annotator.plugins.ExportPlugin;
+import de.unistuttgart.ims.coref.annotator.plugins.ImportPlugin;
 import de.unistuttgart.ims.coref.annotator.plugins.ProcessingPlugin;
 import de.unistuttgart.ims.coref.annotator.plugins.StylePlugin;
+import de.unistuttgart.ims.coref.annotator.plugins.UimaImportPlugin;
 import de.unistuttgart.ims.coref.annotator.profile.Parser;
 import de.unistuttgart.ims.coref.annotator.profile.Profile;
 import de.unistuttgart.ims.coref.annotator.uima.UimaUtil;
@@ -262,6 +265,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		treePopupMenu.add(Annotator.getString(Strings.MENU_EDIT_MENTIONS));
 		treePopupMenu.add(mentionFlagsInTreePopup);
 		treePopupMenu.add(this.actions.mergeMentions);
+		treePopupMenu.add(this.actions.duplicateMentionsAction);
 		treePopupMenu.addSeparator();
 		treePopupMenu.add(Annotator.getString(Strings.MENU_EDIT_ENTITIES));
 		treePopupMenu.add(this.actions.newEntityAction);
@@ -329,6 +333,8 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		controls.add(actions.formGroupAction);
 		controls.add(actions.mergeSelectedEntitiesAction);
 		controls.add(actions.showSearchPanelAction);
+		controls.add(actions.showDocumentStatistics);
+
 		getContentPane().add(controls, BorderLayout.NORTH);
 
 		for (Component comp : controls.getComponents())
@@ -532,6 +538,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		toolsMenu.add(new RenameAllEntitiesAction(this));
 		toolsMenu.add(new RemoveForeignAnnotationsAction(this));
 		toolsMenu.add(new ShowFlagEditor(this));
+		toolsMenu.add(actions.showDocumentStatistics);
 		toolsMenu.addSeparator();
 		// toolsMenu.add(new ShowHistoryAction(this));
 		toolsMenu.add(new ShowLogWindowAction(Annotator.app));
@@ -543,16 +550,13 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		JMenu fileExportMenu = new JMenu(Annotator.getString(Strings.MENU_FILE_EXPORT_AS));
 
 		PluginManager pm = Annotator.app.getPluginManager();
-		for (Class<? extends IOPlugin> pluginClass : pm.getIOPlugins()) {
-			try {
-				IOPlugin plugin = pm.getIOPlugin(pluginClass);
-				if (plugin.getImporter() != null)
-					fileImportMenu.add(new FileImportAction(Annotator.app, plugin));
-				if (plugin.getExporter() != null)
-					fileExportMenu.add(new FileExportAction(this, this, plugin));
-			} catch (ResourceInitializationException e) {
-				Annotator.logger.catching(e);
-			}
+
+		for (ImportPlugin iplugin : pm.getIOPluginObjects().selectInstancesOf(ImportPlugin.class)) {
+			fileImportMenu.add(new FileImportAction(Annotator.app, iplugin));
+		}
+
+		for (ExportPlugin plugin : pm.getIOPluginObjects().selectInstancesOf(ExportPlugin.class)) {
+			fileExportMenu.add(new FileExportAction(this, this, plugin));
 
 		}
 
@@ -646,8 +650,8 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		Annotator.app.close(this);
 	}
 
-	public void loadFile(File file, IOPlugin flavor, String language) {
-		if (flavor instanceof DefaultIOPlugin)
+	public void loadFile(File file, ImportPlugin flavor, String language) {
+		if (flavor instanceof DefaultImportPlugin)
 			this.file = file;
 
 		JCasLoader lai;
@@ -655,28 +659,30 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		setIndeterminateProgress();
 		File profileFile = new File(file.getParentFile(), "profile.xml");
 		final Profile profile = new Parser().getProfileOrNull(profileFile);
-		lai = new JCasLoader(file, flavor, language, jcas -> {
-			this.setJCas(jcas, profile);
-		}, ex -> {
-			String[] options = new String[] { Annotator.getString("message.wrong_file_version.ok"),
-					Annotator.getString("message.wrong_file_version.help") };
-			ImprovedMessageDialog.showMessageDialog(this, Annotator.getString("message.wrong_file_version.title"),
-					ex.getCause().getLocalizedMessage(), options, new BooleanSupplier[] { () -> {
-						return true;
-					}, () -> {
-						try {
-							Desktop.getDesktop().browse(
-									new URI("https://github.com/nilsreiter/CorefAnnotator/wiki/File-format-versions"));
-						} catch (IOException | URISyntaxException e) {
-							Annotator.logger.catching(e);
-						}
-						return true;
-					} });
-			setVisible(false);
-			dispose();
+		if (flavor instanceof UimaImportPlugin) {
+			lai = new JCasLoader(file, (UimaImportPlugin) flavor, language, jcas -> {
+				this.setJCas(jcas, profile);
+			}, ex -> {
+				String[] options = new String[] { Annotator.getString("message.wrong_file_version.ok"),
+						Annotator.getString("message.wrong_file_version.help") };
+				ImprovedMessageDialog.showMessageDialog(this, Annotator.getString("message.wrong_file_version.title"),
+						ex.getCause().getLocalizedMessage(), options, new BooleanSupplier[] { () -> {
+							return true;
+						}, () -> {
+							try {
+								Desktop.getDesktop().browse(new URI(
+										"https://github.com/nilsreiter/CorefAnnotator/wiki/File-format-versions"));
+							} catch (IOException | URISyntaxException e) {
+								Annotator.logger.catching(e);
+							}
+							return true;
+						} });
+				setVisible(false);
+				dispose();
 
-		});
-		lai.execute();
+			});
+			lai.execute();
+		}
 
 	}
 
@@ -1297,6 +1303,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 	}
 
 	class TextViewKeyListener implements KeyListener {
+		@SuppressWarnings("deprecation")
 		@Override
 		public void keyTyped(KeyEvent e) {
 			CoreferenceModel cModel = getDocumentModel().getCoreferenceModel();
@@ -1485,7 +1492,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 
 				// something is selected
 				if (dot != mark) {
-					ImmutableSet<Mention> ms = getDocumentModel().getCoreferenceModel().getMentions(low, high)
+					ImmutableSet<Mention> ms = getDocumentModel().getCoreferenceModel().getMentionsBetween(low, high)
 							.selectInstancesOf(Mention.class);
 					mentions.addAllIterable(ms);
 				}
@@ -1505,6 +1512,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 
 	class TextMouseListener implements MouseListener, MouseMotionListener {
 
+		@SuppressWarnings("deprecation")
 		@Override
 		public void mouseClicked(MouseEvent e) {
 			if (SwingUtilities.isRightMouseButton(e)) {
@@ -1798,6 +1806,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		UndoAction undoAction;
 		AbstractAction setDocumentLanguageAction = new SetLanguageAction(DocumentWindow.this);
 		AbstractAction showSearchPanelAction;
+		AbstractAction showDocumentStatistics = new ShowDocumentStatistics(DocumentWindow.this);
 		SortTree sortByAlpha;
 		SortTree sortByMentions;
 		ToggleEntitySortOrder sortDescending = new ToggleEntitySortOrder(DocumentWindow.this);
@@ -1806,6 +1815,7 @@ public class DocumentWindow extends AbstractTextWindow implements CaretListener,
 		NewEntityAction newEntityAction;
 		RenameEntityAction renameAction;
 		RemoveDuplicatesAction removeDuplicatesAction;
+		DuplicateMentions duplicateMentionsAction = new DuplicateMentions(DocumentWindow.this);
 		EntityStatisticsAction entityStatisticsAction;
 		ViewSetLineNumberStyle lineNumberStyleNone = new ViewSetLineNumberStyle(DocumentWindow.this,
 				LineNumberStyle.NONE);