+++ src/main/java/de/unistuttgart/ims/coref/annotator/Util.java
@@ -1,6 +1,7 @@
 package de.unistuttgart.ims.coref.annotator;
 
 import java.util.Arrays;
+import java.util.Formatter;
 import java.util.Locale;
 
 import javax.swing.tree.TreeModel;
@@ -151,18 +152,22 @@ public class Util {
 		return Util.contains((StringArray) fs.getFeatureValue(feature), flag);
 	}
 
+	@Deprecated
 	public static boolean isGeneric(Entity e) {
 		return Util.contains(e.getFlags(), Constants.ENTITY_FLAG_GENERIC);
 	}
 
+	@Deprecated
 	public static boolean isDifficult(Mention m) {
 		return Util.contains(m.getFlags(), Constants.MENTION_FLAG_DIFFICULT);
 	}
 
+	@Deprecated
 	public static boolean isNonNominal(Mention m) {
 		return Util.contains(m.getFlags(), Constants.MENTION_FLAG_NON_NOMINAL);
 	}
 
+	@Deprecated
 	public static boolean isAmbiguous(Mention m) {
 		return Util.contains(m.getFlags(), Constants.MENTION_FLAG_AMBIGUOUS);
 	}
@@ -290,6 +295,14 @@ public class Util {
 	public static <T extends Enum<?>> T randomEnum(Class<T> clazz) {
 		int x = Constants.RANDOM.nextInt(clazz.getEnumConstants().length);
 		return clazz.getEnumConstants()[x];
+	}
+
+	public static String format(String formatString, Object... objects) {
+		StringBuilder b = new StringBuilder();
+		try (Formatter formatter = new Formatter(b)) {
+			formatter.format(Locale.getDefault(), formatString, objects);
+			return b.toString();
+		}
 	}
 
 }