+++ src/main/java/de/unistuttgart/ims/coref/annotator/plugin/tei/Constants.java
@@ -4,4 +4,5 @@ public class Constants {
 	public static final String PLUGIN_TEI_TEXT_SELECTOR = "PLUGIN_TEI_TEXT_SELECTOR";
 	public static final String PLUGIN_TEI_LANGUAGE = "PLUGIN_TEI_LANGUAGE";
 
+	public static final String NAME = "TEI/XML";
 }