+++ src/main/java/de/unistuttgart/ims/coref/annotator/plugins/DefaultImportPlugin.java
@@ -12,7 +12,6 @@ import org.apache.uima.UIMAException;
 import org.apache.uima.analysis_engine.AnalysisEngineDescription;
 import org.apache.uima.cas.impl.XmiCasDeserializer;
 import org.apache.uima.collection.CollectionReaderDescription;
-import org.apache.uima.fit.component.NoOpAnnotator;
 import org.apache.uima.fit.factory.AggregateBuilder;
 import org.apache.uima.fit.factory.AnalysisEngineFactory;
 import org.apache.uima.fit.factory.CollectionReaderFactory;
@@ -28,7 +27,7 @@ import de.unistuttgart.ims.coref.annotator.ExtensionFilters;
 import de.unistuttgart.ims.coref.annotator.FileFilters;
 import javafx.stage.FileChooser.ExtensionFilter;
 
-public final class DefaultIOPlugin extends AbstractXmiPlugin implements DirectFileIOPlugin {
+public final class DefaultImportPlugin extends AbstractImportPlugin implements DirectFileIOPlugin, UimaImportPlugin {
 
 	File lastFile;
 
@@ -54,11 +53,6 @@ public final class DefaultIOPlugin extends AbstractXmiPlugin implements DirectFi
 		AggregateBuilder b1 = new AggregateBuilder();
 		b1.add(AnalysisEngineFactory.createEngineDescription(CheckLoadability.class));
 		return b1.createAggregateDescription();
-	}
-
-	@Override
-	public AnalysisEngineDescription getExporter() throws ResourceInitializationException {
-		return AnalysisEngineFactory.createEngineDescription(NoOpAnnotator.class);
 	}
 
 	@Override
@@ -105,6 +99,11 @@ public final class DefaultIOPlugin extends AbstractXmiPlugin implements DirectFi
 	@Override
 	public Ikon getIkon() {
 		return MaterialDesign.MDI_FILE_EXPORT;
+	}
+
+	@Override
+	public String getSuffix() {
+		return ".xmi";
 	}
 
 }