+++ src/main/java/de/unistuttgart/ims/coref/annotator/plugins/ExportPlugin.java
@@ -1,0 +1,9 @@
+package de.unistuttgart.ims.coref.annotator.plugins;
+
+import java.io.File;
+import java.util.function.Consumer;
+
+public interface ExportPlugin extends IOPlugin {
+	Consumer<File> getPostExportAction();
+
+}